# Sched refactor to only query necessary attribs