<!--- Please review your changes in preview mode -->
<!--- Provide a general summary of your changes in the Title above -->

#### Describe Bug or Feature
<!--- Describe the problem, ideally from the customer's viewpoint  -->
Right now, scheduler queries all attributes of jobs when querying the universe every cycle, but only uses a limited set from them. So, there's time and memory wasted in packaging & sending unnecessary data over the wire, so we should refactor the code to only query the attributes that it needs.

#### Describe Your Change
<!--- Say how you fixed the problem.  Please describe your code changes in detail for reviewer -->
Just added a list of attributes to query in the selstat request.

#### Link to Design Doc
<!--- If there is a design, link to it here: **[project documentation area](https://pbspro.atlassian.net/wiki/display/PD)** -->


#### Attach Test and Valgrind Logs/Output
<!--- Please attach your test log output from running the test you added (or from existing tests that cover your changes) -->
<!--- Don't forget to run Valgrind if appropriate and attach the resulting logs -->
Results of the performance test:
[ravi@pbspro ~]$ grep "Average sched cycle length for last 100 cycles:" before.log 
2019-06-27 06:53:16,532 INFO     Average sched cycle length for last 100 cycles: 6.18350519657
[ravi@pbspro ~]$ grep "Average sched cycle length for last 100 cycles:" after.log 
2019-06-26 18:36:44,384 INFO     Average sched cycle length for last 100 cycles: 4.9587498045

Even though the difference is just over a second, it accumulates over scheduling cycles. So, after 3600 cycles, we'll have saved an hour of time in scheduling, so it is actually quite significant.

Full logs:
[before.log](https://github.com/PBSPro/pbspro/files/3333476/before.log)
[after.log](https://github.com/PBSPro/pbspro/files/3333477/after.log)



<!--- Pull Request Guidelines: [Pull Request Guidelines](https://pbspro.atlassian.net/wiki/spaces/DG/pages/1187348483/Pull+Request+Guidelines) -->