@agrawalravi90 2019-07-18
@bhroam and @arungrover since the performance boost from this isn't significant unless we have 300k+ jobs in the queue, a performance test for this change isn't practical. So, I've removed the one I'd written, which wasn't showing much benefit because there were only 10k jobs in the queue.

Please resume review for this and let me know if changes look ok, sorry for the delay on this.